差不多就做完了

+++ app/src/main/java/org/chengjian/java/feidian/collectdata/beans/CitySellRent.java
@@ -59,9 +59,9 @@ public class CitySellRent {
 
     private String detail;              //详细说明
 
-    private float longitude;        //经度
+    private double longitude;        //经度
 
-    private float latitude;         //维度
+    private double latitude;         //维度
 
     private String researcher;      //调查人
 
@@ -263,19 +263,19 @@ public class CitySellRent {
         this.detail = detail;
     }
 
-    public float getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(float longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
 
-    public float getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(float latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
